requirements.txt
discord.py
flask
python-dotenv
wikipedia